requirements.txt
datasets>=2.19.0
openai>=1.30.0
pyyaml>=6.0
gitpython>=3.1.40
swebench>=2.0.0
